Refactoring <stv-stacked-bar-chart>
