--- src/js/axis_label.js
@@ -1,7 +1,0 @@
-/**
- * src/js/axis_label.js
- */
-// function selectAxisLabelFontSize(selector) {
-//   var val = document.querySelector('#axisLabelFontSize').value;
-//   getChartEl(selector).axisLabelFontSize = val;
-// }
