--- src/js/bar_width.js
@@ -1,8 +1,0 @@
-/**
- * src/js/bar_width.js
- * Toggling the max bar width for bar charts
- */
-// function selectMaxBarWidth(selector) {
-//   var val = document.querySelector('#maxBarWidth').value;
-//   getChartEl(selector).maxBarWidth = val;
-// }
