--- src/js/color_scheme.js
@@ -1,8 +1,0 @@
-/**
- * src/js/color_scheme.js
- * Toggling the applied color scheme
- */
-// function setColorScheme(selector) {
-//   var val = document.querySelector('#colorScheme').value;
-//   getChartEl(selector).colorScheme = val;
-// }
