--- src/js/gridline.js
@@ -1,10 +1,0 @@
-/**
- * src/js/gridline.js
- * Toggle grid lines
- */
-// var gridlines = false;
-
-// function toggleGridlines(selector) {
-//   gridlines = !gridlines;
-//   getChartEl(selector).gridlines = gridlines;
-// }
