--- src/js/interpolation.js
@@ -1,7 +1,0 @@
-/**
- * src/js/interpolation.js
- */
-// function selectInterpolation(selector) {
-//   var val = document.querySelector('#interpolation').value;
-//   getChartEl(selector).interpolation = val;
-// }
