--- src/js/inverse.js
@@ -1,16 +1,0 @@
-/**
- * src/js/inverse.js
- * inverting the colors
- */
-// var inverse = false;
-
-// function toggleInverse(selector) {
-//   inverse = !inverse;
-//   getChartEl(selector).inverse = inverse;
-
-//   if (inverse) {
-//     document.querySelector('div.chart').classList.add('inverse')
-//   } else {
-//     document.querySelector('div.chart').classList.remove('inverse')
-//   }
-// }
