--- src/js/legend.js
@@ -1,9 +1,0 @@
-/**
- * src/js/legend.js
- */
-// var showLegend = false;
-
-// function toggleLegend(selector) {
-//   showLegend = !showLegend;
-//   getChartEl(selector).legend = showLegend;
-// }
