--- src/js/margins.js
@@ -1,9 +1,0 @@
-/**
- * src/js/margins.js
- * Controlling top, right, bottom, and left margins
- */
-function setMargin(marginAttribute, selector) {
-  var val = document.querySelector('#' + marginAttribute).value;
-  var el = getChartEl(selector)
-  el[marginAttribute] = val;
-}
